Implemented protocol and transport.

+++ FastTrackIpc/FastTrackIpc/ITransport.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstddef>
+
 class ITransport {
   public:
     virtual ~ITransport() {}
 
-    virtual void readBytes(char * data, int size) = 0;
+    virtual void readBytes(char * data, size_t size) = 0;
 
-    virtual void writeBytes(const char * data, int size) = 0;
+    virtual void writeBytes(const char * data, size_t size) = 0;
 };
